auth part 2&3

+++ controller/note.js
@@ -3,6 +3,7 @@ const Note = require("../model/Note");
 exports.create = (req, res, next) => {
   const note = new Note({
     ...req.body,
+    _userId: req.auth.userId,
   });
   note
     .save()
@@ -11,7 +12,7 @@ exports.create = (req, res, next) => {
 };
 
 exports.getAll = (req, res, next) => {
-  Note.find()
+  Note.find({ userId: req.auth.userId })
     .then((notes) => res.status(200).json(notes))
     .catch((err) => res.status(400).json({ err }));
 };
@@ -23,17 +24,38 @@ exports.getById = (req, res, next) => {
 };
 
 exports.updateById = (req, res, next) => {
-  Note.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => {
-      return Note.findOne({ _id: req.params.id })
-        .then((note) => res.status(200).json(note))
-        .catch((error) => res.status(404).json({ error }));
+  Note.findOne({ _id: req.params.id })
+    .then((note) => {
+      if (note.userId != req.auth.userId) {
+        res.status(401).json({ message: "Unauthorized" });
+      } else {
+        Note.updateOne(
+          { _id: req.params.id },
+          { ...req.body, _id: req.params.id }
+        )
+          .then(
+            Note.findOne({ _id: req.params.id })
+              .then((note) => res.status(200).json(note))
+              .catch((err) => res.status(400).json({ err }))
+          )
+          .catch((error) => res.status(401).json({ error }));
+      }
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((err) => res.status(400).json({ err }));
 };
 
 exports.deleteById = (req, res, next) => {
-  Note.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé" }))
-    .catch((error) => res.status(400).json({ error }));
+  Note.findOne({ _id: req.params.id })
+    .then((note) => {
+      if (note.userId != req.auth.userId) {
+        res.status(401).json({ message: "Unauthorized" });
+      } else {
+        Note.deleteOne({ _id: req.params.id })
+          .then(() => {
+            res.status(200).json({ message: "Note deleted" });
+          })
+          .catch((error) => res.status(401).json({ error }));
+      }
+    })
+    .catch((err) => req.status(400).json({ err }));
 };
